% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abi.R
\name{abi}
\alias{abi}
\title{tree-sitter ABI version}
\usage{
abi()
}
\value{
A single integer.
}
\description{
\code{abi()} returns the ABI version that \code{parser.c} was generated with. This is
used to verify that the grammar is compatible with the tree-sitter C library
embedded within the treesitter package.
}
\examples{
abi()
}
