% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binding.R
\name{language}
\alias{language}
\title{tree-sitter language for R}
\usage{
language()
}
\value{
An external pointer.
}
\description{
\code{language()} returns an external pointer to the tree-sitter language
object for R.
}
\examples{
language()
}
