% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesitter.r-package.R
\docType{package}
\name{treesitter.r-package}
\alias{treesitter.r}
\alias{treesitter.r-package}
\title{treesitter.r: 'R' Grammar for 'Tree-Sitter'}
\description{
Provides bindings to an 'R' grammar for 'Tree-sitter', to be used alongside the 'treesitter' package. 'Tree-sitter' builds concrete syntax trees for source files of any language, and can efficiently update those syntax trees as the source file is edited.
}
\author{
\strong{Maintainer}: Davis Vaughan \email{davis@posit.co}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
