% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesitterr-package.R
\docType{package}
\name{treesitterr-package}
\alias{treesitterr}
\alias{treesitterr-package}
\title{treesitterr: R Grammar for tree-sitter}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Davis Vaughan \email{davis@posit.co}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
