% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language.R
\name{language}
\alias{language}
\title{tree-sitter language for R}
\usage{
language()
}
\value{
A \code{tree_sitter_language} object.
}
\description{
\code{language()} returns a \code{tree_sitter_language} object for R for use with the
treesitter package.
}
\examples{
language()
}
