% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesitterr-package.R
\docType{package}
\name{treesitterr-package}
\alias{treesitterr}
\alias{treesitterr-package}
\title{treesitterr: R Grammar for tree-sitter}
\description{
Provides bindings to an R grammar for tree-sitter, to be used alongside the 'treesitter' package. Tree-sitter builds concrete syntax trees for source files of any language, and can efficiently update those syntax trees as the source file is edited.
}
\author{
\strong{Maintainer}: Davis Vaughan \email{davis@posit.co}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
